Fixing test for signal interrupt exception
